Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I have the 3G version on the Vega so the phone signal indicator is useful. This needs to be switchable in some way so that if radio is present it will still display.

I will submit a patch to get 3G working shortly.

Patch-set: 1