vega: Add RIL and GPS for devices with Ericsson F3307
Shuttle offer the p10an01 with an optional Ericsson F3307 3G card which includes GPS.

init.rc already tries to start rild with the correct library, but the library is missing. Include the library, plus include the GPS libs.

mbm_config should stop rild from trying to start if no card is present, at the moment it loops looking for the lib.

The Shuttle rom's also include libs for a Huawei 3G card, but I don't know if that was ever actually sold in the p10an01. Some markets may get that 3G card.

Change-Id: I11cf47c36c173d5360bb5779941238a51adab5f9