Update patch set 1
Patch Set 1:

The Huawei modem uses different libs. I wasn't sure if it was actually sold anywhere with that modem, so I didn't really look at it.

If we have someone with the Huawei modem that can do some testing I could try to get it to use the correct libs if the hardware is present, but without someone to test it I would only be guessing what will work. Do you know what model the Huawei card is? Maybe I can find one to test.

Patch-set: 1