Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Don't merge this for now until we get the Huawei bits included.

That looks like it also partially works for Huawei USB dongles. I have a Huawei E160E dongle, so I might try to get that working.

I found a version of the Huawei card which will work on the networks here at a reasonable price, so I might buy it for my laptop and can use it to test the vega.

Patch-set: 1